EM/PowerUI: Screen will be turned on at low battery status
Summary:
When Screen is off, then low battery popup is not visible to user.
With this patch we make the screen turn on just before popup and
it will remain in this state for 3 sec.
Also, if in silent-mode, then vibra will be turned on for 0.5 sec.

Orig-Change-Id: If665f47f27d933f59ce5f5895dac91091e7f2fdf

Change-Id: I618c8544229adb71870010f26407022450137a98
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Signed-off-by: Jian Luo <jian.luo@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 52107

Test Plan: it builds, it works.

Reviewers: randall.rushing, synergy

Reviewed By: synergy

Differential Revision: http://review.codefirex.com/D413